Changed GameStats into a module, updated StatTracker and tests
