+++ lib/stat_tracker.rb
@@ -1,5 +1,6 @@
 class StatTracker < StatSorter
 
+  include GameStats
   include LeagueStats
 
   attr_reader :files,
@@ -11,7 +12,7 @@ class StatTracker < StatSorter
     @files = files
     @games = convert_files.games
     @teams = convert_files.teams
-    @game_stats = compile_stats[:games]
+    # @game_stats = compile_stats[:games]
     @team_stats = compile_stats[:teams]
     # @league_stats = compile_stats[:league]
     update_stats
@@ -27,7 +28,7 @@ class StatTracker < StatSorter
 
   def compile_stats
     {
-    games: GameStats.new(@games),
+    # games: GameStats.new(@games),
     teams: TeamStats.new(@teams),
     # league: LeagueStats.new(@games, @teams)
     }
@@ -35,53 +36,53 @@ class StatTracker < StatSorter
 
 # Game Statistics
 # ======================================
-  def highest_total_score
-    game_stats.max_score
-  end
-
-  def lowest_total_score
-    game_stats.min_score
-  end
-
-  def biggest_blowout
-    game_stats.blowout
-  end
-
-  def most_popular_venue
-    game_stats.venue_with_most_games
-  end
-
-  def least_popular_venue
-    game_stats.venue_with_fewest_games
-  end
-
+  # def highest_total_score
+  #   game_stats.max_score
+  # end
+  #
+  # def lowest_total_score
+  #   game_stats.min_score
+  # end
+  #
+  # def biggest_blowout
+  #   game_stats.blowout
+  # end
+  #
+  # def most_popular_venue
+  #   game_stats.venue_with_most_games
+  # end
+  #
+  # def least_popular_venue
+  #   game_stats.venue_with_fewest_games
+  # end
+  #
   def percentage_home_wins
-    game_stats.wins_percentage("home")
+    wins_percentage("home")
   end
 
   def percentage_visitor_wins
-    game_stats.wins_percentage("away")
+    wins_percentage("away")
   end
-
-  def season_with_most_games
-    game_stats.season_with_most_games
-  end
-
-  def season_with_fewest_games
-    game_stats.season_with_fewest_games
-  end
-
-  def count_of_games_by_season
-    game_stats.list_season_games
-  end
-
-  def average_goals_per_game
-    game_stats.average_game_goals
-  end
-
-  def average_goals_by_season
-    game_stats.average_goals_by_season
-  end
+  #
+  # def season_with_most_games
+  #   game_stats.season_with_most_games
+  # end
+  #
+  # def season_with_fewest_games
+  #   game_stats.season_with_fewest_games
+  # end
+  #
+  # def count_of_games_by_season
+  #   game_stats.list_season_games
+  # end
+  #
+  # def average_goals_per_game
+  #   game_stats.average_game_goals
+  # end
+  #
+  # def average_goals_by_season
+  #   game_stats.average_goals_by_season
+  # end
 
   # League Statistics
   # ======================================
